+++ packages/runtime/src/widgets/Form/index.ts
@@ -4,3 +4,4 @@ export * from "./Checkbox";
 export * from "./Radio";
 export * from "./Dropdown";
 export * from "./MultiSelect";
+export * from "./Date/Date";
